+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/upgrade/SimpleUpgradeItem.java
@@ -63,7 +63,6 @@ public final class SimpleUpgradeItem extends AbstractUpgradeItem {
         );
     }
 
-    // TODO: more than 1 regulator in exporter and importer?
     public static SimpleUpgradeItem autocraftingUpgrade() {
         return new SimpleUpgradeItem(
             RefinedStorageApi.INSTANCE.getUpgradeRegistry(),